fix(auth, roles)

+++ src/post/post.controller.ts
@@ -15,12 +15,17 @@ import { PostsService } from './post.service';
 import { CreatePostDto, UpdatePostsDto } from './dto/register-post.dto';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 import { Posts } from './dto/post';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { Role } from '../auth/enums/rol.enum';
+import { Auth } from '../auth/decorators/auth.decorator';
 
+@ApiTags('Posts')
 @Controller('posts')
 export class PostController {
   constructor(private readonly postsService: PostsService) {}
   @Post('')
-  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  @Auth(Role.ADMIN)
   async register(@Body() dto: CreatePostDto, @Req() request): Promise<void> {
     const author = request.user.userId;
     const newDto = { ...dto, author };
@@ -52,33 +57,38 @@ export class PostController {
   @Put(':id')
   @UseGuards(JwtAuthGuard)
   async update(
-    @Param('id') postId: string,
+    @Req() request,
+    @Param('id') postsId: string,
     @Body() dto: UpdatePostsDto,
   ): Promise<Posts> {
-    const user = await this.postsService.update(postId, dto);
+    const postsDB = await this.postsService.findById(postsId);
+    const isAdmin = request.user.isAdmin;
+    const isYourSelft = postsDB.author == request.user.userId ? true : false;
+    if (!(isAdmin || isYourSelft)) {
+      throw new NotFoundException('Forbidden resource');
+    }
+    const user = await this.postsService.update(postsId, dto);
     if (!user) {
       throw new NotFoundException('Posts not found');
     }
     return user;
   }
 
+  @ApiBearerAuth()
   @Delete(':id')
   @UseGuards(JwtAuthGuard)
-  async delete(@Param('id') postsId: string): Promise<boolean> {
+  async delete(@Req() request, @Param('id') postsId: string): Promise<boolean> {
+    const postsDB = await this.postsService.findById(postsId);
+    const isAdmin = request.user.isAdmin;
+    const isYourSelft = postsDB.author == request.user.userId ? true : false;
+    if (!(isAdmin || isYourSelft)) {
+      throw new NotFoundException('Forbidden resource');
+    }
     const posts = await this.postsService.delete(postsId);
     if (!posts) {
       throw new NotFoundException('User not found');
     }
     return posts;
-  }
-
-  @Get('search')
-  async findByCriteria(@Query() query: Record<string, any>): Promise<Posts[]> {
-    const user = await this.postsService.searchByCriteria(query);
-    if (!user) {
-      throw new NotFoundException('Posts not found');
-    }
-    return user;
   }
 
   @Get('user/:id')
@@ -89,6 +99,17 @@ export class PostController {
     }
     return posts;
   }
+
+  @ApiTags('Filter')
+  @Get('search')
+  async findByCriteria(@Query() query: Record<string, any>): Promise<Posts[]> {
+    const user = await this.postsService.searchByCriteria(query);
+    if (!user) {
+      throw new NotFoundException('Posts not found');
+    }
+    return user;
+  }
+  @ApiTags('Filter')
   @Get('filter/')
   async findByCategories(
     @Query('categories') categories: string,
